[Backport] Security bug 905509 (2/13)
Backport of original patch by Antoine Labour <piman@chromium.org>:
Have CommandBuffer::CreateTransferBuffer take a uint32_t instead of
size_t

Because client and service may be of different bitness, offsets in
command buffers have to fit in a uint32_t, effectively limiting transfer
buffers to 4GB.
Make this clear in CommandBuffer::CreateTransferBuffer by taking a
uin32_t instead of a size_t (fixing callers as appropriate), avoiding
potential security issues with silent clamping.

Bug: 905509
Reviewed-on: https://chromium-review.googlesource.com/c/1396132

+++ chromium/gpu/command_buffer/client/mock_transfer_buffer.cc
@@ -79,7 +79,7 @@ void* MockTransferBuffer::AllocUpTo(unsigned int size,
   // reallocated.
   actual_buffer_index_ = (actual_buffer_index_ + 1) % kNumBuffers;
 
-  size = std::min(static_cast<size_t>(size), MaxTransferBufferSize());
+  size = std::min(size, MaxTransferBufferSize());
   if (actual_offset_ + size > size_) {
     actual_offset_ = result_size_;
   }
@@ -129,7 +129,7 @@ unsigned int MockTransferBuffer::GetFragmentedFreeSize() const {
 
 void MockTransferBuffer::ShrinkLastBlock(unsigned int new_size) {}
 
-size_t MockTransferBuffer::MaxTransferBufferSize() {
+uint32_t MockTransferBuffer::MaxTransferBufferSize() {
   return size_ - result_size_;
 }
 
@@ -143,7 +143,7 @@ bool MockTransferBuffer::InSync() {
 }
 
 MockTransferBuffer::ExpectedMemoryInfo MockTransferBuffer::GetExpectedMemory(
-    size_t size) {
+    uint32_t size) {
   ExpectedMemoryInfo mem;
   mem.offset = AllocateExpectedTransferBuffer(size);
   mem.id = GetExpectedTransferBufferId();
@@ -153,7 +153,7 @@ MockTransferBuffer::ExpectedMemoryInfo MockTransferBuffer::GetExpectedMemory(
 }
 
 MockTransferBuffer::ExpectedMemoryInfo
-MockTransferBuffer::GetExpectedResultMemory(size_t size) {
+MockTransferBuffer::GetExpectedResultMemory(uint32_t size) {
   ExpectedMemoryInfo mem;
   mem.offset = GetExpectedResultBufferOffset();
   mem.id = GetExpectedResultBufferId();
@@ -162,7 +162,7 @@ MockTransferBuffer::GetExpectedResultMemory(size_t size) {
   return mem;
 }
 
-uint32_t MockTransferBuffer::AllocateExpectedTransferBuffer(size_t size) {
+uint32_t MockTransferBuffer::AllocateExpectedTransferBuffer(uint32_t size) {
   EXPECT_LE(size, MaxTransferBufferSize());
 
   // Toggle which buffer we get each time to simulate the buffer being
@@ -180,7 +180,7 @@ uint32_t MockTransferBuffer::AllocateExpectedTransferBuffer(size_t size) {
 }
 
 void* MockTransferBuffer::GetExpectedTransferAddressFromOffset(uint32_t offset,
-                                                               size_t size) {
+                                                               uint32_t size) {
   EXPECT_GE(offset, expected_buffer_index_ * alignment_);
   EXPECT_LE(offset + size, size_ + expected_buffer_index_ * alignment_);
   return expected_buffer() + offset;
