+++ chromium/gpu/ipc/in_process_command_buffer.cc
@@ -799,7 +799,7 @@ void InProcessCommandBuffer::SetGetBufferOnGpuThread(
 }
 
 scoped_refptr<Buffer> InProcessCommandBuffer::CreateTransferBuffer(
-    size_t size,
+    uint32_t size,
     int32_t* id) {
   CheckSequencedThread();
   base::AutoLock lock(command_buffer_lock_);
